% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwqs_main-revised.R
\name{bwqs_plot}
\alias{bwqs_plot}
\title{Plot function for Bayesian Weighted Quantile Sum regression models}
\usage{
bwqs_plot(
  stanfit,
  parms = NULL,
  cri_level = 0.95,
  point_color = "black",
  line_color = "black",
  shape = 21,
  size = 5
)
}
\arguments{
\item{stanfit}{the object coming from \code{bwqs()} function.}

\item{parms}{a vector with the name of parameters that we want to plot possible value are
\code{'beta0'}, \code{'beta1'}, \code{'W'}, \code{'delta'} (components of covariates) and
\code{'sigma'} (if \code{'gaussian'} family is selected).}

\item{cri_level}{An \code{integer} containing the credible level for the interval. This number must
be between 0 and 1.}

\item{point_color}{color of point that indicates the value for the fitted parameters (default \code{black}).}

\item{line_color}{color of line that indicates the credible interval (default \code{black}).}

\item{shape}{symbols to indicate the value of parameters.}

\item{size}{dimension of \code{shape}.}
}
\description{
This function provide a plot of the parameters with a credible interval. The paramenters of
interest could be manually selected and also the level of credible interval.
}
\details{
For examples of this function see example of \code{bwqs} function
}
