% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwqs_main-revised.R
\name{bwqs_r}
\alias{bwqs_r}
\title{Fitting Bayesian Weighted Quantile Sum regression models}
\usage{
bwqs_r(
  formula,
  mix_name,
  cluster_name,
  data,
  q,
  Dalp = NULL,
  chains = 1,
  iter = 1000,
  thin = 3,
  seed = 2019,
  start_value = NULL,
  c_int = c(0.025, 0.975),
  family = "gaussian"
)
}
\arguments{
\item{formula}{Object of class \code{formula} specifying the relationship between the outcome and the
covariates of the model not involved in the mixture variable. If the model has no covariates specify
\code{y ~ NULL}.}

\item{mix_name}{A character vector listing the variables contributing to a mixture effect.}

\item{cluster_name}{A character string that specifiy which is the column of the dataset which
contains the group number. Note that the \code{cluster_name} should be numeric, strings and factors
are not allowed}

\item{data}{The \code{data.frame} containing the variables (covariates and elements of the mixture)
to be included in the model.}

\item{q}{An \code{integer} to specify how mixture variables will be ranked, e.g. in quartiles
(\code{q = 4}), deciles (\code{q = 10}), or percentiles (\code{q = 100}). If \code{q = NULL} then
the values of the mixture variables are taken (these must be standardized or the domain must be the same).}

\item{Dalp}{A \code{vector} containing the parameters of the Dirichlet distribution of the weights, the number
of the elements of the vector has to be equal to the number of chemicals. If \code{Dalp = NULL}
the domain is explored uniformly.}

\item{chains}{An \code{integer} to specify the number of chain in Hamiltonian Monte Carlo algorithm.
Default value \code{chains = 1}.}

\item{iter}{An \code{integer} to specify the lenght of chain in Hamiltonian Monte Carlo algorithm.
Default value \code{iter = 10000}.}

\item{thin}{An \code{integer} to specify the thinning parameter in Hamiltonian Monte Carlo algorithm.}

\item{seed}{An \code{integer} value to fix the seed. If \code{seed = NULL} the seed are randomly choosen.}

\item{start_value}{A \code{vector} containing the initial value of the prior distribution,
if it is equal to NULL random values are chosen.}

\item{c_int}{A \code{vector} of two elements to specify the credible intervals for parameters, for 95\% credible
interval \code{c_int = c(0.025,0.975)} (default).}

\item{family}{A \code{string} to specify the type of outcome. With the current implementation  the
possible values are only continuous - "gaussian" (default).}
}
\value{
\code{bwqs} returns a list with two argument:
\item{fit}{An \code{S4} object with all details of the Hamiltonian Monte Carlo, all the extractions
from the posterior distribution and all values of the parameters}
\item{summary_fit}{Table with the statistics of the parameters: mean, standard error of the mean,
standard deviation, lower and upper values for the credible interval (with credible level specified
by \code{c_int}), n_eff and Rhat. For further details see https://cran.r-project.org/web/packages/rstan/rstan.pdf}
}
\description{
Fits Random Bayesian Weighted Quantile Sum (BWQS) regressions for continuous outcomes. This model
provides estimation for the mixture composition and overall effect of the mixture across different
groups on the outcomes using bayesian framework.
}
\details{
The function \code{bwqs} uses the package \code{rstan} which allows the connection with STAN,
a specific software, written in C++ for bayesian inference, for further information see https://mc-stan.org/.
}
\author{
Nicolo Foppa Pedretti, Elena Colicino
}
