% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwqs_main-revised.R
\name{bwqs_waic}
\alias{bwqs_waic}
\title{Compute of Watanabe-Akaike Information Criterion}
\usage{
bwqs_waic(stanfit)
}
\arguments{
\item{stanfit}{Object \code{fit} is returned from the \code{bwqs} function.}
}
\value{
The function returns a list with:
\item{waic}{value of total WAIC for the model}
\item{elpd_waic}{expected log pointwise predictive density for a new dataset}
\item{p_waic}{the estimated effective number of parameters}
\item{elpd_loo}{approximation to the actual out-of-sample prediction error under the model}
\item{p_loo}{effective number of parameters as
the bias adjustment corresponding to the overfitting inherent}
}
\description{
This function compute the WAIC and LOO given a posterior distribution.
The values of the logarithmic likelihood are extracted from Stan object
of the bwqs function.
}
\details{
This function is implemented based on the formulas given
in http://www.stat.columbia.edu/~gelman/research/unpublished/waic_stan.
For examples of this function see example of \code{bwqs} function
}
