% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwqs_main-revised.R
\name{quantile_split}
\alias{quantile_split}
\title{Quantile function for BWQS}
\usage{
quantile_split(data, mix_name = mix_name, q, shift = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} where the columns have the name of
the mixture components.}

\item{mix_name}{A character list of components that have been
considered in the mixture.}

\item{q}{An \code{integer} to specify how mixture variables will be standardize, e.g. in quartiles
(\code{q = 4}), deciles (\code{q = 10}), or percentiles (\code{q = 100}).}

\item{shift}{If \code{SHITF = TRUE} the quantiles are counted from 0, otherwise 1.}
}
\value{
A \code{data.frame} with quantiles for element of the list only.
}
\description{
This function allows quantile splitting for BWQS model.
}
\details{
For examples of this function see example of \code{bwqs} function
}
