% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwqs_main-revised.R
\name{quantile_split_2}
\alias{quantile_split_2}
\title{Quantile function for BWQS with continuous data}
\usage{
quantile_split_2(data, mix_name = mix_name, q)
}
\arguments{
\item{data}{A \code{data.frame} where the columns have the name of
the mixture components.}

\item{mix_name}{A character list of components that have been
considered in the mixture.}

\item{q}{An \code{integer} to specify how mixture variables will be standardize, e.g. in quartiles
(\code{q = 4}), deciles (\code{q = 10}), or percentiles (\code{q = 100}).}
}
\value{
A \code{data.frame} with specified quantiles for element of the list only.
}
\description{
This function allows quantile splitting (empirical cumulative function)
for BWQS model.
}
\details{
For examples of this function see example of \code{bwqs} function
}
